[chromiumos-overlay] Separate autotest frontend from autotest-server
To speed up the process of emerging autotest-server, a new ebuild,
autotest-server-web-frontend, has been added that handles the
compilation of the frontend javascript.

BUG=chromium:479319
TEST=Tested making minor changes to the autotest server and frontend
code, and then deploying the changes separately on a moblab and
verifying that the changes only appeared when the appropriate
package was deployed.

Change-Id: Ie7389bf4ae478f3483ae21f743127b25f092330a
Reviewed-on: https://chromium-review.googlesource.com/268440
Tested-by: Matthew Sartori <msartori@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Matthew Sartori <msartori@chromium.org>