initramfs: Build the fitpicker utility into the loader kernel if needed.
Add a dependency on the fitpicker ebuild so it will be available, and let the
Makefiles know whether to include it. We need it to prepare a FIT style kernel
for kexec when chaining through a non-vboot kernel.

BUG=brillo:893
TEST=Built the initramfs for the loader kernel on peach_pit and saw that the
fitpicker utility was built as well. Built it for lumpy and saw that it was
unaffected.

Change-Id: Iaafac0239396c22c5e564dbcaf547618f30d0d34
Reviewed-on: https://chromium-review.googlesource.com/270007
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>