python: patch semaphore.c to fix pid namespace bug
Currently python2.7 tries to create semaphores as a file like
"/dev/shm/mp<pid>-<counter>".  This assumes that there is only one
process per pid, which is not the case with pid namespacing which is
used to run builds in parallel.

The fix is to add another suffix, the current time's microseconds, which
should very likely be unique.  Furthermore, we retry up to 100 times
until there is no collision.

The change to dev-install is necessary because of crbug.com/489895.

BUG=chromium:481223
TEST=No more of Python's unit tests failed; a sqlite3 test was broken.

Change-Id: Ib80f2b9d26d6c7309372742bbd16ffd28cb56e25
Reviewed-on: https://chromium-review.googlesource.com/271417
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Paul Hobbs <phobbs@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>