chromeos-seabios: fix build tool generation
The ebuild runs make with the compiler set for the target when it tries
to generate tools that'll be running on the build system (e.g. kconfig).
This works by accident (our chroot has 32bit & 64bit libs, and they are
mostly compatible with the target libs), but is fundamentally wrong.

BUG=chromium:488360
TEST=precq passes

Change-Id: Ic940eb6cffa757de827237db3d3739627c9e091c
Reviewed-on: https://chromium-review.googlesource.com/271634
Trybot-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>