sys-boot: Dynamically decide which stages to extract, deprecate USE=vboot2
As we're adding more and more verified firmware components, this patch
deduplicates a bit of the copy&pasted code for them in the ebuilds. It
removes the explicit reliance on USE flags when deciding which stages to
extract and instead goes with a permissive "whatever is there" approach.
This works even for ramstage and romstage since cros_bundle_firmware
just picks whatever it needs (as dictated by the FMAP) from its blobs
array, and doesn't care if there are any extra entries. (Also add
support for the new bl31 component type that will be needed by ARM TF,
although the FMAP changes for individual boards using it have not been
committed yet.)

With these changes and the removal of a few other no longer needed
pieces, we can completely retire the vboot2 USE flag (since the decision
to use vboot2 originates in coreboot Kconfigs anyway).

BUG=None
TEST=Compiled Oak and Jerry, booted Falco.

Change-Id: I88d6fb247631b87d2fae5f330bdfddf66789a405
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/272298
Reviewed-by: Aaron Durbin <adurbin@chromium.org>