dhcpcd: port patches to version 6.8.2
While there, add new patches to clean up the old RPC interface to allow
easy integration of new RPC mechanisms in the future and enable hook
scripts when no RPC is configured (which allows the dhcpcd to be compiled
and used the same way as upstream).

Note: Patches are applied in the order that non-ChromiumOS specific patches
are applied first before the OS specific patch. So those patches doesn't
contain any OS specific changes, and can be upstreamed without much
change if we chose to.

Note: RPC support is not integrated with IPv6 yet.

BUG=chromium:486122
TEST=Run DHCP client autotests
TEST=Run wifi_matfunc suite
CQ-DEPEND=CL:272615

Change-Id: I43bc3be98a50c693135be20e635594a15a8f49ad
Reviewed-on: https://chromium-review.googlesource.com/272614
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Zeping Qiu <zqiu@chromium.org>
Commit-Queue: Zeping Qiu <zqiu@chromium.org>