cros-kernel: add .scmversion file for 9999 builds
For cros_workon (9999) builds, add a .scmversion file to the output
directory. This will prevent the git sha from being added to the kernel
version and is useful for local developer builds.

BUG=chrome-os-partner:40556
TEST=Couple of test cases:
1. for cros_workon kernel build, verify that the kernel version
is of the form 3.18.0 and doesn't and doesn't include git hash in the
kernel version.
2. for non-cros_workon kernel builds, if CONFIG_LOCALVERSION_AUTO is
enabled, the kernel version should include the git sha.

Change-Id: I08afc0e22ecc62b26e9d75eec2868b1dea94f607
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/273282
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>