chromeos-chrome: filter out -clang-syntax
We plan to add -clang-syntax to the make.conf of some boards to make
sure the chromeos tree is clang clean. chromeos-chrome itself should be
clang clean, and we have asan bots that build chrome with clang. So we
do not need to check it again.
On the other hand, to build chome with clang, one needs to set clang as
the compiler in the configure so that it can set/disable approperiate
flags for clang. Using clang to compile chrome without specifying
USE='clang' would trigger compile warnings/errors like unknown compilier
flags and so on. So I think it is better to filter out this flag for
chromeos-chrome.

BUG=None
TEST=CXXFLAGS=-clang-syntax emerge-falco chromeos-chrome
     clang is not invoked.

Change-Id: I173a33b27b7e932ce2584176768b172bb2e372ab
Reviewed-on: https://chromium-review.googlesource.com/273531
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>