Land initial sandbox-spec-gen ebuild
We're changing 'container' to 'sandbox', so we need to install the
right new protos and change the name of the package that installs
them. Unfortunately, we can't currently land a change that requires
a new ebuild added in that change to init the SDK.

Land the ebuild first, then use it in a later change.

BUG=brillo:1028
TEST=None

Change-Id: I7ec294f74a02853ac5f13fe51ebd383dae48124d
Reviewed-on: https://chromium-review.googlesource.com/273677
Trybot-Ready: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>