Add a param to doappid to specify the devicetype
This will add DEVICETYPE=<devicetype> to /etc/lsb-release. This allows
Chrome to accurately determine the type of device it is running on.

This parameter is currently optional until the existing doappid usages
are modified to use it. The parameter will become required in the near
future.

More context on this change can be found in the following review:
https://chromium-review.googlesource.com/#/c/273714/.

BUG=chromium:341356
TEST=Verify that /etc/lsb-release contains a string with
DEVICETYPE=<something>.
Change-Id: I3c91ae3bb7a0923fcf213fb1e3467cb5a33e3c3e
Reviewed-on: https://chromium-review.googlesource.com/273689
Reviewed-by: Bertrand Simonnet <bsimonnet@chromium.org>
Commit-Queue: Jacob Dufault <jdufault@chromium.org>
Tested-by: Jacob Dufault <jdufault@chromium.org>