Add policy_EditBookmarksEnabled to autotest-chrome-9999.ebuild.
CL 234901 added the Enterprise Policy Test framework for developers
and test engineers to easily write tests to verify the effect of policies
on the behavior and appearance of Chrome OS. The CL included a dummy test,
policy_EditBookmarksEnabled, which exercises this framework.

A CL will be provided in the next few days that will turn Pinocchio into a
real test.

BUG=none
TEST=Run test_that <IP> policy_EditBookmarksEnabled

Change-Id: I7cd4918cb432752eaac61a3e5e2514c38e9ae99c
Reviewed-on: https://chromium-review.googlesource.com/273954
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Scott Cunningham <scunningham@chromium.org>
Reviewed-by: Scott Cunningham <scunningham@chromium.org>
Commit-Queue: Scott Cunningham <scunningham@chromium.org>