seperate the removal of '-clang-syntax' from use_gcc
Some packages that can build with clang can not build with
'-clang-syntax'. It is because the are either using CFLAGS
to host compiler or confusing the configeration. So we need
to filter out -clang-syntax for these packages but still
enable them to build with clang on clang bot.

BUG=none
TEST=CC=x86_64-cros-linux-gnu-clang emerge-falco syslinux
     CFLAGS='-clang-syntax' emerge-falco syslinux

Change-Id: I463a3f374f5662a7f795b62564aff279ea1cc730
Reviewed-on: https://chromium-review.googlesource.com/274081
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>