make.conf: punt $ARCH-target files
At this point, the amd64 & arm ones are the same as the generic target.
The x86 one has custom MARCH_TUNE settings, but those are being moved to
the few boards that need it.  So we can punt these and always use the one
generic file.

BUG=chromium:489641
TEST=precq passes
CQ-DEPEND=CL:273968

Change-Id: Ic0f428a9169d6d0ac97230bea03db190fa48d9e3
Reviewed-on: https://chromium-review.googlesource.com/274754
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Nam Nguyen <namnguyen@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>