binutils: force to use gcc.
Our binutils cannot be built with clang. Upstream has a fix for it, but
it may take time to backport to our tree. As binutils is the only
package left to fix for the amd64-generic-llvm build, we force it to use
gcc for now.

BUG=chromium:477433
TEST=CC=x86_64-cros-linux-gnu-clang CXX=x86_64-cros-linux-gnu-clang++
     emerge-falco binutils

Change-Id: Icf49ae91b9cddd877ebb58b879db4361c45ba1b5
Reviewed-on: https://chromium-review.googlesource.com/275276
Reviewed-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>