dhcpcd: Remove errant 0 return in dhcp_envoption
The "return 0" while processing embedded options in
dhcp_envoption() is dangerous, since it under-reports
the data it will actually have written to the env
array.  Callers use the returned value to set the
allocation size of the env array by calling this method
with a NULL env pointer.  Thus, with the function as it
is, there is a real possibility that dhcp_envoption may
write past the end of env in a subsequent call.

Cases where the optlen has been zero have been benign
so far, so instead of an early return (with a bogus and
perilously ignored value), this CL simply breaks out
of the embedded-options loop and continues to the
encapsulated loop, and proceeds to accurately return the
correct allocation size to the caller.

BUG=chromium:500112
TEST=Manual: Connect with Messi-V/ALT3100 modem
TEST=Note the message "encountered zero-length opt" in /var/log/net.log
TEST=Note LTE connects successfully
TEST=CL:280380 adds automated test "network_DhcpFQDN" that replicates the issue

Change-Id: I41fe2a090de9f4df977d15061e8965e0bf9d1f55
Reviewed-on: https://chromium-review.googlesource.com/278362
Reviewed-by: Rebecca Silberstein <silberst@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>