Fix openssl clang compilation error.
"openssl-1.0.2" is not buildable using clang, the clang-assembler does
not recognize some arm assembly.

One of them is "ldrplb", this must be re-written to "ldrbpl" for clang-assmebler.
Similarly, "vmov.32" must be re-writtent to "vmov". (see bug)

Upstream CL 313e6ec1, which aims at providing support for
Apple Ios contains fix for this. This CL is part of the upstream CL -
313e6ec1.

BUG=chromium:502312
TEST=emerged openssl correctly using clang and gcc, cbuildbot falco-release daisy-release

Change-Id: I21a8e49a216b338eb16d6e33223dd684aa7c36d7
Reviewed-on: https://chromium-review.googlesource.com/280605
Reviewed-by: Yunlian Jiang <yunlian@chromium.org>
Commit-Queue: Han Shen <shenhan@chromium.org>
Tested-by: Han Shen <shenhan@chromium.org>