dhcpcd: Take upstream fix for dhcp_envoption
The "return 0" while processing embedded options in
dhcp_envoption() is dangerous, since it under-reports
the data it will actually have written to the env
array.  Callers use the returned value to set the
allocation size of the env array by calling this method
with a NULL env pointer.  Thus, with the function as it
is, there is a real possibility that dhcp_envoption may
write past the end of env in a subsequent call.

This CL replaes the fix from CL:278362 with the version
that is now upstream.

BUG=chromium:500112
TEST=Manual: Connect with Messi-V/ALT3100 modem
TEST=Note LTE connects successfully
TEST=CL:280380 adds automated test "network_DhcpFQDN" that replicates the issue

Change-Id: Iedae7f646c239b801196ef484de995523ffbf48b
Reviewed-on: https://chromium-review.googlesource.com/281000
Reviewed-by: Zeping Qiu <zqiu@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>