sshd: Create README for starting sshd on base images.
Currently sshd is installed on all images types, but only dev and test
images start sshd on boot, and it can be difficult to determine how to
start sshd on a base image.

This CL adds a README file to give a few examples of ways to start sshd
on a base image if needed.

BUG=chromium:467210
TEST=Followed the steps in the README to verify their accuracy.
TEST=Built and installed image, README file is in the expected location.

Change-Id: Idd52b5ac38b4a99b80fd973381faf54c531730a1
Reviewed-on: https://chromium-review.googlesource.com/281052
Reviewed-by: David Pursell <dpursell@chromium.org>
Commit-Queue: David Pursell <dpursell@chromium.org>
Trybot-Ready: David Pursell <dpursell@chromium.org>
Tested-by: David Pursell <dpursell@chromium.org>