cros-kernel2: Compress ARM64 kernel binaries
This patch adds compression support to ARM64 kernels. Since those don't
have a built-in decompressor, we'll compress them externally and note
this in the appropriate place of the FIT image descriptor. Our firmware
supports both LZMA and LZ4, but LZ4 seems to have much better boot time
implications on our current ARM64 SoCs (about 20 times faster
decompression speed at about 1.5 times the compressed size), so let's go
with that for now.

A small complication arises from the fact that the ARM64 kernel wants
to be relocated to a specific offset within a 2MB-aligned block for
whatever reason, and that it communicates this request through a header
at the top of the kernel binary. We don't want to load the image,
decompress it, read that header and then memcpy() the decompressed image
again, so we extract this offset beforehand and encode it in the "load"
property of the FIT image (which is supposed to be an absolute load
address, but, meh... close enough, and as far as I can tell U-Boot
ignores it for kernel_noload images anyway).

Also add app-arch/lz4 to target-chromium-os-sdk RDEPENDS to ensure its
available for use by this ebuild (is that the right way to do this?).

CQ-DEPEND=CL:282752,CL:281805,CL:286221
BUG=chrome-os-partner:32184
TEST=Built and booted LZMA compressed Oak kernel image. Noted that it
works but takes about 800(!)ms of decompression time. Scrambled to hack
an LZ4 algorithm into coreboot and noted a much more manageable ~80ms.

Change-Id: I55056a95e2d5fd543bec1b0c7eec3324f4faa812
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/281851
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>