update_engine: Remove unused libpcre dependency.
libpcre was not used in update_engine as far as I can tell from the git
history, and the dependency was removed from the SConstruct file in 2013
in CL:66381. This patch just removes the dependency from the ebuild.

BUG=None
TEST=emerge-link update_engine

Change-Id: I4f1d28c16adde277f40993fe92e53f1b6372770a
Reviewed-on: https://chromium-review.googlesource.com/283928
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>