chromeos-sshd-init: Add -w option to invokation of `iptables` command
When multiple processes use `iptables` to modify the firewall, the
command grabs an exclusive lock for the table being modified. If the
lock cannot be obtained (another instance of iptables is running),
the current instance fails with an error.

By adding -w we make it wait for the other lock to be released before
proceeding.

BUG=brillo:1240
TEST=`FEATURES=test emerge-gizmo chromeos-sshd-init`
     Ran numerous reboot tests with validation of `iptables -S`
     to make sure port 22 is open after reboot completes.

Change-Id: I75835be4435a1d0655c59f83396ad2cb10167079
Reviewed-on: https://chromium-review.googlesource.com/285551
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Vitaly Buka <vitalybuka@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>