trunks: use a dedicated USE flag to enable FTDI support
Until now SPI over FTDI support was enabled when building for the
host. But now want to move FTDI support from the client to the daemon,
which can be used only on Chrome OS devices. To be able to do it let's
introduce a dedicated USE flag (ftdi_tpm).

BUG=chromium:498861
TEST=with the appropriate changes in the daemon in place verified that
     FTDI interface can be used on a samus.

Change-Id: If41bf9de70f070a4284553cf56cd581e2af0f111
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/285534
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>