chromeos-ec: Add support for split EC image build.
With the changes to the build infrastructure, the EC objects are now
built separately for both RO and RW. This is in preparation for a shared
objects library. The ebuild needed to have its paths updated to account
for this.

BRANCH=None
BUG=None
TEST=`emerge-rambi chromeos-ec` and verified that all the right firmware
images were present in the /build/rambi/firmware directory.
TEST=`emerge-rush_ryu chromeos-bootimage` and verified that it completed
successfully.

CQ-DEPEND=CL:274079

Change-Id: I25b3095558290f65d9b04877ab922abb8e7dee43
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/285934
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>