Update harfbuzz to 1.0.1
In addition to getting the latest and greatest,
this is to sync harfbuzz in CrOS with that in Chromium on other
platforms where a bundled harfbuzz in Chromium tree is used.

The upstream does not have the ebuild for 1.0.1. The upstream's
latest ebuild 0.9.41 was downloaded and renamed. Besides, KEYWORDS
is set to '*' to enable it on all platforms.

BUG=chromium:517603
TEST=emerge-{x86-alex,amd64-generic,daisy} harfbuzz succeeds.
TEST=cbuildbot chromiumos_sdk succeeds.
TEST=cbuildbot x86-alex VMtest passes.

Change-Id: Ib26025333d535a5b9eaef9e2183b605755605940
Reviewed-on: https://chromium-review.googlesource.com/291781
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Jungshik Shin <jshin@chromium.org>
Commit-Queue: Jungshik Shin <jshin@chromium.org>