chromeos-fonts: fallback to qemu for native runs
If the native chroot failed, try running it through qemu.  When the target
code is built for the target board, it might use insns that are not usable
on the build system.  We can't do this all the time as qemu is much slower
than native runs (like 40seconds vs 15seconds).

BUG=chromium:473900
TEST=`emerge-samus chromeos-fonts` still works

Change-Id: Ib986c721a8b0c8836940ac17ea73524bafa30972
Reviewed-on: https://chromium-review.googlesource.com/294923
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jungshik Shin <jshin@chromium.org>