udev-init-scripts: punt!
We already disable this package for the target which means it's only
used in the sdk.  Tweak the USE flags to also cover the sdk and punt
the package from the system.  It only installs one script which is
used when booting, but we don't do that in the sdk, so we can punt it.

BUG=chromium:455336
TEST=`cbuildbot chromiumos-sdk` passes

Change-Id: I10f5bb773d61a194df1f0451d25544f41da4b88e
Reviewed-on: https://chromium-review.googlesource.com/296001
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>