chromeos-init: Give the proper name for "halt" when using s3halt
When we were using s3halt we had the name of the "halt" upstart script
as "s3halt".  This didn't prevent the script from running (since it
still had "start on runlevel 0") but it did stop other things from
running at shutdown time since "pre-shutdown.conf" had "start on
starting halt".

That resulted in the system thinking that an unclean shutdown
happened.  You could see this because "mmcblk0" would show that it was
being "recovered" at reboot and also you could see that Chrome would
complain that you didn't shutdown properly.  With this change both
symptoms are gone.

Note that this breaks wake from BT, but that will need to be addressed
separately.

BUG=chrome-os-partner:44679
TEST=No more errors after shutdown; can still wake from USB

Change-Id: I88aa21fe5b8ffa1d4d1bb1799ccb518a1851a9a3
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/296422
Reviewed-by: Dan Erat <derat@chromium.org>