cros-kernel2: Add Rockchip graphics stuff to fallback config
It's nice to make graphics work.  Add a few misc configs for that.  Note
that the MALI stuff is only relevant for Heiko's "somewhat stable"
branch, but that's still useful.

I'm not 100% sure that all of these are necessary, but they seem to work
OK.

BUG=None
TEST=Boot up and see graphics

Change-Id: I6e72178df48250cb13af45fc316e9ddde8710739
Reviewed-on: https://chromium-review.googlesource.com/296839
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>