libchromeos: Install libpolicy from the top level directory.
libpolicy is not installed under the chromeos/ directory, so we moved the
chromeos/policy directory to the top level in the source code. This
patch updates the ebuild to install it from the right location and
bumps the stable version to the latest code.

Pulled 2 new changes from platform/external/libchromeos: (ebuild: chromeos-base/libchromeos)
  532fcf2 libpolicy: Fix unittest paths.
  3c8d4ab Add libpolicy stub-implementation to Android.mk

BUG=b:23555535
TEST=emerge-link libchromeos; `mm` in aosp.

Change-Id: I423f17471b8cec03e1324603f0ad76e1192a9377
Reviewed-on: https://chromium-review.googlesource.com/296835
Commit-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Alex Deymo <deymo@chromium.org>