wpa_supplicant: UPSTREAM: do not add blacklists on normal disconnection
There are number of cases where wpa_supplicant requests the current
connection to be disconnected before starting a new operation. Such
cases do not really indicate that there was an error in connecting or a
disconnection initiated by the AP, so do not add a temporary blacklist
entry in such sequences.

BUG=chromium:527481
TEST=run autotests on a veyron_jaq chromebook

Change-Id: I5b1ee33cc660f87b737bb9f966189fbe2bdc3b56
Reviewed-on: https://chromium-review.googlesource.com/298605
Commit-Ready: Ningyuan Wang <nywang@chromium.org>
Tested-by: Ningyuan Wang <nywang@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>