Make DBus client libraries depend on the generator.
Building DBus client libraries for the host requires the DBus bindings
generator installed on the host. When building for the target board
this is not an issue since all the host ebuilds are installed before
we build any of the ebuilds for the target board. This patch makes the
dependency explicit when building for the host.

BUG=chromium:530265
TEST=./update_chroot

Change-Id: I983251734dece1fb4f98f32042ccf50988e0a1f1
Reviewed-on: https://chromium-review.googlesource.com/298901
Commit-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>