mesa-img: Surfaceless: flush when switching contex
Make sure we issue a flush when unbinding a context, otherwise
switching to a new context may leave the old one unrendered.

(Previously we wouldn't do a flush if there was no surface bound.
This optimisation is not correct in a surfaceless environment.)

Also do some other minor fixups to the ebuild as recommended in
previous reviews.

BUG=chrome-os-partner:44961
TEST=WebGL field, hopalong and aquarium
CQ-DEPEND=CL:*230856

Change-Id: I8a5ef976879eec9b45cb98edb324ff08dd939c12
Signed-off-by: Rufus Hamade <rufus.hamade@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/299842
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>