Filter errata fix options for chromeos-chrome build.
We had a previous CL that adds '-Wl,--fix-cortex-a53-843419' option to
the aarch64 target linker, howerver, this option was inproperly
propagated to ChromeSDK builder in building a host binary, and caused a
build error. Considering Chrome in chromeos is still 32-bit application,
I filter this option for a workaround.

BUG=chromium:532532
TEST=cbuildbot oak-full

Change-Id: I4d9d44a15452eb164b4982df3fdec0b80ec2c87e
Reviewed-on: https://chromium-review.googlesource.com/299791
Commit-Ready: Han Shen <shenhan@chromium.org>
Tested-by: Han Shen <shenhan@chromium.org>
Reviewed-by: Han Shen <shenhan@chromium.org>