dhcpcd: ChromiumOS: Ensure gateway probe is broadcast
The arp_new() method does not always return a freshly
initialized arp_state structure.  For example, if
dhcp_probe_gw() is called while a probe from
start_unicast_arp() is still running, the arp_state
object from the unicast ARP call will be returned by
arp_new().  This can happen when a DHCP renew succeeds
while a unicast "GatewayArp" is still in progress.

In the above scenario we need to make sure the dest_hwlen
field of the arp_state structure is zero.  This signals a
broadcast ARP, and will ensure that the callbacks that are
invoked by the ARP state machine know which case we are in.

While here, add a small optimization that skips a
broadcast gateway ARP if a unicast ARP succeeds.

BUG=chromium:531042
TEST=Reconnect to WiFi network where ARP fails

Change-Id: I0505804d07960561e164b6612528605d1f19e1ab
Reviewed-on: https://chromium-review.googlesource.com/300343
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>