brillo-base/libsparse: Add ebuild with libsparse tools.
libsparse is part of the AOSP platform/system.core repo. It consists on
a library and a set of command line tools to manipulate Android sparse
images.

This patch creates a new cros-workonable ebuild based on the local copy
of platform/system/core, blacklisted. The Makefile used to build is not
the Android.mk included in the source code. Instead, this ebuild uses
its own Makefile, based on the deleted Gentoo version of
dev-utils/libsparse.

This ebuild is now included as a dependency of cros-devutils, used in
brillo_update_payload.

BUG=b:23599483
TEST=sudo emerge libsparse; emerge-daisy_spring libsparse
TEST=simg2img system.img system.raw

Change-Id: I95e63d42d425faccc809a3c9f9ed56ba516e35ef
Reviewed-on: https://chromium-review.googlesource.com/300654
Commit-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>