chromeos-chrome.ebuild: put gsutil binary files inside /var/cache/chromeos-chrome
depot_tools/gsutil.py which is used by some hooks, tries to download gsutil
from the network and save it save it on disk. By default it save that into
a subdirectory inside depot_tools.
This change will make it cache binaries under /var/cache/chromeos-chrome -
a directory which is guaranteed to be writable.

BUG=chromium:452497
TEST="ebuild-amd64-generic `equery-amd64-generic w chromeos-chrome` configure"
  finishes without errors and caches gsutil in /var/cache/chromeos-chrome.

Change-Id: Ie6003af059189f548c064826d5b8fe0d039cd0cc
Reviewed-on: https://chromium-review.googlesource.com/301751
Commit-Ready: Andrey Ulanov <andreyu@google.com>
Tested-by: Andrey Ulanov <andreyu@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>