mesa-img: fix segfault when calling ClientWaitSync
Mesa has a bug where it tries to use the current context when
eglClientWaitSyncKHR is called.  If that function is called from
a thread without a current context then it would segfault.

This change adds a patch to fix Mesa so that it handles this
scenareo correctly.
 - If there is no current context, don't attempt to use it
 - If EGL_SYNC_FLUSH_COMMANDS_BIT_KHR is set and there is no
   current context, then ignore this bit.

It is then up to the HW driver to do the correct thing if
client_wait_sync is invoked without a context.  (The IMG driver
already does the correct thing.)

BUG=chrome-os-partner:45537
TEST=Boot to UI, run WebGL conformance 1.0.2 video tests.

Change-Id: I80a0d2d5fe84dfd724ccb1b8daa1eece44d434b7
Signed-off-by: Rufus Hamade <rufus.hamade@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/302310
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>