dev-lang/go: a new ebuild for Go 1.5.1.
This ebuild brings Go 1.5.1 to ChromeOS (both host and cross compilers).
Details on testing and known issues can be found at http://go/Go1.5-On-ChromeOS.

With Go 1.5, compiler and runtime are now written entirely in Go, and as such,
we need an older installation of Go to bootstrap and build Go 1.5 from source.
This ebuild uses Go 1.4.2 to bootstrap Go 1.5.1.

BUG=chromium:372937
TEST='emerge dev-lang/go' and ran compiler and standard library tests on host.
TEST='emerge cross-x86_64-cros-linux-gnu/go' and ran compiler and standard library tests on a panther device.
TEST='emerge cross-i686-pc-linux-gnu/go' and ran compiler and standard library tests on a x86-zgb device.
TEST='emerge cross-armv7a-cros-linux-gnueabi/go' and ran compiler and standard library tests on a daisy device.

Change-Id: I5183d757bff1ea6c7ae397893909e92be78d4b0c
Reviewed-on: https://chromium-review.googlesource.com/303494
Commit-Ready: Rahul Chaudhry <rahulchaudhry@chromium.org>
Tested-by: Rahul Chaudhry <rahulchaudhry@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>