New dev-libs/xz-embedded ebuild.
xz-embedded is small XZ decompressor implemented by the same authors
that created xz-utils. The code was originally developed for the linux
kernel but it can be easily used in other projects.

This patch creates an ebuild for the latest xz-embedded release that
installs a statically linked library with the pre-compiled code. There
is no API nor ABI stability guarantee between versions of XZ Embedded.
Nevertheless, the code didn't change in the last few years.

The library headers and the current API are different than the liblzma
library, so they can be installed together and used independently. The
embedded version is suitable for client-side decompression since the
compiled code is smaller. An equivalent .so file in amd64 is about
19 KiB, compared to the 143 KiB for liblzma.

BUG=b:23604708
TEST=emerge-link xz-embedded; emerge-daisy_spring xz-embedded

Change-Id: I2f4bc731d96a7512b7c8d84d191ba413976cefeb
Reviewed-on: https://chromium-review.googlesource.com/303292
Commit-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>