chromeos-base/update_engine: Depend on xz-embedded.
dev-libs/xz-embedded is used in the update_engine daemon to apply
REPLACE_XZ operations. Note that we also DEPEND on xz-utils to build
delta_generator and RDEPEND on xz-utils for the host to run it, but not
on the target as it is only used for creating payloads.

CQ-DEPEND=CL:303292
BUG=b:23604708
TEST=`emerge-link update_engine` pulls in xz-embedded.

Change-Id: I78843ec3ac807f14caa67133ea1ed4132ae5fb89
Reviewed-on: https://chromium-review.googlesource.com/303219
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>