skylake boards: update vboot firmware indicies
These indicies are the correct ones for skylake boards. I
checked in a change to coreboot that honored the location of
the FSP in the RW slots. However, I had tested it locally with
a similar change as this. Obviously it worked. Not so much w/o
the config change.

BUG=chrome-os-partner:46583
BRANCH=None
TEST=Built and booted glados.

Change-Id: Ie7036e961c39f276a8ea07d9a1fc954852cf5666
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/306720
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>