Move libmnl to chromiumos-overlay and fix clang build error.
The source code in libmnl tries to change the visibility of
functions after they have been defined.  GCC allows this, but clang
does not.  The result is that, when built with clang, most functions
are 'hidden' and get stripped out altogether.  Then anything that
tries to use these library functions fails to build.

This CL creates a version of the package in chromiumos-overlay
(it was in portage stable) and applies an epatch that fixes the
source problems for now.

The libmnl-function-attributes.patch below was also submitted to
the netfilter-devel list, in an attempt to get the patch
taken/reviewed upstream:

http://marc.info/?l=netfilter-devel&m=144606523930462&w=2

BUG=chromium:548786
TEST=Successfully emerged net-libs/libmnl and sys-apps/iproute2
with clang, with these changes.

Change-Id: Ib470266a82395ab73db11e05890d8ab264eda8e4
Reviewed-on: https://chromium-review.googlesource.com/309605
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>