freetype: upgrade to 2.6.2 with one line patch.
Upgraded media-libs/freetype to version 2.6.2 on amd64, arm, x86
with one local patch to turn back ON 'stem darkening' for CFF.

This is done to make it safe for merging to the branch with
the benefit of security hardening in 2.6.2.
We may turn stem darkening off again along with hint-slight for
our CFF fonts in fontconfig in the trunk after merge.

To have this local patch, freetype is now in chromiumos-overaly.

BUG=chromium:551028
TEST=`emerge-{BOARD} freetype` on x86-generic, daisy, link
TEST=`cbuildbot --remote -g 315881 chromiumos-sdk
      x86-alex-release daisy-release`

Change-Id: I5e3556b14524c3df14181e708ee8bd10fd53834d
Reviewed-on: https://chromium-review.googlesource.com/315881
Commit-Ready: Jungshik Shin <jshin@chromium.org>
Tested-by: Jungshik Shin <jshin@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>