mesa: update surface/sampler_view destruction fix for softpipe
update the 10.3-state_tracker-gallium-fix-crash-with-st_renderbuffer patch
to include a similar fix for the softpipe driver in order to avoid the
crash on surface/sampler_view destruction when the context is gone.

BUG=chrome-os-partner:48215
TEST=emerge-smaug-cheets mesa
run Chrome on Smaug and no longer see the crash at startup in
swrast.

Change-Id: I752afb1c15ff17687ccb57f0841a13d5c530acd1
Reviewed-on: https://chromium-review.googlesource.com/316915
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>