openssl: fix llvm builds (part 2 of 2).
This CL manually picks up upstream sources that are guided by
"__APPLE__" for sha256-armv4.pl, this is only possible after a bug in
llvm 'sub' is fixed in CL 320302.

This CL also turns on llvm build for the package.

The bug entry for llvm-built openssl fails to boot on arm board is here
- https://llvm.org/bugs/show_bug.cgi?id=24346

CQ-DEPEND=CL:320302
BUG=None
TEST=build openssl using patched llvm, and got a good image. &&
cbuildbot on daisy-release, lumpy-release.

Change-Id: I565ba8aea04108460a55ec11cde98ba0df4eee49
Reviewed-on: https://chromium-review.googlesource.com/320461
Commit-Ready: Han Shen <shenhan@chromium.org>
Tested-by: Han Shen <shenhan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>