GCC sysroot_wrapper: Remove temporary workaround.
The Clang compiler has been updated to handle the flag
'-march=armv8-a+crc', so the temporary workaround in the
sysroot_wrapper.hardened file, which looked for and removed
that option, is no longer needed and can be removed.

BUG=chromium:554632
TEST=Tested the Clang in the latest ChromeOS with this change
and it recognized the option and handled it correctly.

Change-Id: Id1a3c4ce311befd6c1dec2aeaa0bc3539be5ead6
Reviewed-on: https://chromium-review.googlesource.com/320893
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Yunlian Jiang <yunlian@chromium.org>